+++ packages/@sourcegraph/extension-api-types/src/url.d.ts
@@ -1,0 +1,2 @@
+export type ReadonlySearchParams = Readonly<Omit<URLSearchParams, 'append' | 'set' | 'delete'>>
+export type ReadonlyURL = Readonly<Omit<URL, 'searchParams'> & { searchParams: ReadonlySearchParams }>